fix(langchain_google): Add web platform compatibility for HttpClientAuthProvider
Use conditional exports to isolate dart:io dependencies, following the same pattern as googleai_dart's FilesResource.

- On IO platforms: Uses the full implementation with googleapis_auth
- On web/WASM: Uses a stub that throws UnsupportedError with helpful message

Fixes #812
